+++ src/server/index.js
@@ -1,1 +1,4 @@
-module.exports = {};
+const registerBabel = require("./register-babel");
+module.exports = {
+  registerBabel
+};