+++ CHANGELOG.md
@@ -1,14 +1,39 @@
+# 9/13/2019
+
+#### Unreleased
+
+### Creates
+
+- Add `DomElement.md`
+- Add `PopupWindow.md`
+- Add `register-babel.md`
+
+### Updates
+
+- Update `README.md`
+- Update `DomElement` functions, add `prepend`
+- Update `DomElement.spec.js` test cases
+- Update `PopupWindow` styles and default values
+
+### Removals
+
+- Remove `clis/utils.js`
+
 # 9/12/2019
 
 #### [1.0.1](https://github.com/1846689910/cicero/commit/d98c32b2c8f2195eeffc5272f7c9ff6275e4a80c)
 
 ### Creates
+
 - Added `DomElement` test cases
 - Added `LICENSE`
+
 ### Updates
+
 - Updated `DomElement`
 - Updated `PopupWindow`
 - mod `package.json`
+
 ### Removals
 
 # 9/10/2019